Support nullish coalescing operator in TypeScript lexer (#1529)
TypeScript has added support for the nullish coalescing operator `??` as
of version 3.7. This commit adds support by prepending a rule to the
`:root` state of lexers that extend the TypeScript common module.
When JavaScript adopts the operator, the rule should be moved into the
`:root` state of the JavaScript lexer and removed from the TypeScript
common module.