Support scope resolution operator in C++ lexer (#1523)
The C++ lexer inherits the majority of its rules from the C lexer.
While this approach reduces duplication, it can cause issues where
assumptions that are correct in C, cause issues in C++.

One such issue occurs when the scope resolution operator, `::`, is used
in a case statement. The C lexer contains a rule that treats colons in
a case statement as representing the 'end' of that state. While the
assumption that a colon would not otherwise occur in a case statement
is correct in C, that is not the case in C++. The scope resolution
operator uses two colons and it can occur within a case statement.

This commit creates a new `:case` state in the C++ lexer that uses a
negative lookahead to avoid matching a scope resolution operator. A
rule matching the scope resolution operator is also prepended to the
`:statements` state.