Improve support for single quotes in Haskell lexer (#1524)
Haskell uses matching single quotes to express character literals (e.g.
`'A'`). However, the unmatched single quote character, `'`, is also
used for various purposes (e.g. as part of a variable name like `x'` or
in the name of a promoted type like `'Bar`). Currently, Rouge's Haskell
lexer permits the use of an unmatched single quote by having separate
rules for naming and then a general rule of lower precedence for when
the lexer encounters `'`.

A better approach would be to only treat `'` as being part of a
character literal if the lexer detects that there is a matching `'`
within an appropriate number of characters. This commit does that and
then creates a more general rule for matching names.