Fix incorrect predicate usage in PowerShell lexer (#1536)
The PowerShell lexer includes rules that use the existence of a nil
value in a regular expression to decide whether to push new states onto
the stack. The predicates for these rules were written incorrectly.
This commit removes one of the predicates (which wasn't necessary) and
uses `Object#nil?` in the other to match correctly.