Remove docker_ file glob from Docker lexer (#1550)
In addition to adding the `*.Dockerfile` file glob, #1059 added a file
glob for `Dockerfile_*` to address #1070. However, as noted by #1539,
this causes issues for files that begin with `dockerfile_` but that are
not Dockerfiles.

Adding the `Dockerfile_*` file glob was an error and this commit
removes it. It is a non-standard name and is not recognised by other
syntax highlighting libraries (e.g. Pygments, Chroma, Linguist). The
correct fix for the issue that precipitated #1070 is to use a
`.gitattributes` file.