Fix character escaping in Perl lexer (#1549)
In single- and double-quoted strings, the Perl lexer currently
tokenises `\` as an error if it is not followed by a character in part
of a recognised escape sequence. This is a bug. Perl accepts the use of
`\` in single- and double-quoted strings even if it is not part of a
valid escape sequence.

This commit permits the use of `\` in single-quoted and double-quoted
strings as well as increasing the range of escape sequences that are
recognised in double-quoted strings.