Rewrite PHP lexer to support use statements, function declarations and type declarations (#1489)
This commit involves a large-scale rewrite of the PHP lexer with the aim
of improving the level of semantic detail. Partly this is achieved by
the use of various heuristics that, while not always correct, do hold
for the majority of cases. In particular, the lexer has improved
support for `use` statements, function declarations and type
declarations.

Co-authored-by: julp <julp@users.noreply.github.com>