Handle % like / in Ruby lexer (#1563)
The Ruby lexer only treats `%` consistently as the modulo operator if it
is followed by whitespace. This is not consistent with Ruby itself which
treats this use of `%` as the modulo operator:

    if board[i/w][i%s] == 'O'
      region << i
    end

This is a difficult problem to solve because Ruby's parser categorises
`%` based on the variables which are in scope (which is not something
Rouge can do). This commit improves the handling of `%` by treating it
in a similar way to `/`.