apex: fix invalid use of String#casecmp (#1596)
In ruby, this always returns a number, which... is always truthy. Yes,
even 0 is truthy in Ruby. (Also casecmp is 0 when they're equal, so it
would have needed negation...? idk.) This seems like the original intent
of the code, and makes the highlighting look wayyyy better. We were
mismatching colours on parentheses ><

Co-authored-by: http://jneen.net/ <jneen@jneen.net>