Improve handling of data structure literals in PowerShell lexer (#1595)
The PowerShell lexer includes a hash table state as a special case and
is not tested on more complex tables (e.g. including comments and
code). This commit factors out the hash table state into a more generic
`:expr` state that can handle more complex structures.

This commit also reduces the overall level of highlighting. This 
improves the visibility of the more important syntax.