Ensure word break follows keywords in Kotlin lexer (#1621)
Rouge will currently match identifiers in the Kotlin lexer that begin
with certain keywords (such as `super`). This is because of a change in
#1496 that added a rule to support labels. That change should have
updated the rule to end with a `\b` so as to ensure that identifiers
that begin with the listed keywords are followed by a word break.
This commit fixes that error.