Revert empty patterns in Ruby lexer (#1624)
The maintenance changes in #1548 resulted in updates to two rules in
the Ruby lexer, one regarding regular expression flags and the other
regarding heredoc strings. However, these patterns should not have been
changed. The change in #1548 was to prevent rules that matched an empty
string unless they contained a 'predicate' of some sort. Both rules in
the Ruby lexer contained such a predicate and so should not have been
changed. This commit fixes that error.