Support floor division operator in Crystal lexer (#1639)
Currently, Crystal's floor division operator (`//`) is not lexed
properly when `:root` is the topmost state in the stack. Instead, Rouge
lexs it as the regular division operator (`/`) followed by a regex
literal.

This commit adds `//` to the rule that matches `**` and other repeated
operator characters. It also reorders some of these operators to ensure
that they are matched correctly.

Co-authored-by: Michael Camilleri <mike@inqk.net>