Add support for HTTP/2 responses to HTTP lexer (#1864)
The existing HTTP lexer parses HTTP/2 requests, but not responses marked with HTTP/2. (However, it would accept HTTP/2.0.)

There is no specification for the textual representation of HTTP/2 traffic (as far as I can tell), but some tools generate response headers with only HTTP/2, which fail to be processed correctly by Rouge without this change.