Replace usage of `cgi` for ruby 3.5 compatibility
In Ruby 3.5, most of the cgi gem is removed. That includes `CGI.parse`, which `rouge` uses.

But, we can subsistute it with `URI.decode_www_form`.
The data is returned in a slightly different format (array of arrays instead of a hash),
so do some processing to get back into the expected structure.

https://bugs.ruby-lang.org/issues/21258